requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.0
jinja2==3.1.2
bcrypt==4.0.1
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
mailjet-rest==1.3.4

